+++ CHANGELOG.md
@@ -2,6 +2,50 @@
 All notable changes to this project will be documented in this file.
 
 ---
+## [2.4.0](https://github.com/VoIPGRID/VialerSIPLib/tree/2.4.0) (01/18/2017)
+Released on Wednesday, January 18, 2017.
+
+#### Added
+* Added custom ringtone if file is present (#49)[https://github.com/VoIPGRID/VialerSIPLib/pull/49]
+
+#### Fixed
+* Remove if check on error pointer when creating VSLCall (#49)[https://github.com/VoIPGRID/VialerSIPLib/pull/49]
+
+
+## [2.3.0](https://github.com/VoIPGRID/VialerSIPLib/tree/2.3.0) (01/18/2017)
+Released on Wednesday, January 18, 2017.
+
+#### Added
+* Show stats after call in example app (#46)[https://github.com/VoIPGRID/VialerSIPLib/pull/46]
+* Added log callback so that implementing app can get logs (#47)[https://github.com/VoIPGRID/VialerSIPLib/pull/47]
+
+
+## [2.2.0](https://github.com/VoIPGRID/VialerSIPLib/tree/2.2.0) (01/13/2017)
+Released on Friday, January 13, 2017.
+
+#### Added
+* Enable TCP or UDP switch in the Example app (#42)[https://github.com/VoIPGRID/VialerSIPLib/pull/42]
+
+#### Fixed
+* VSLAudiocontroller forward declaration is replaced by importing header. (#37)[https://github.com/VoIPGRID/VialerSIPLib/pull/37]
+* Start monitoring network changes after the call started. (#38)[https://github.com/VoIPGRID/VialerSIPLib/pull/38]
+* Update VIA headers when sending reINVITES. (#38)[https://github.com/VoIPGRID/VialerSIPLib/pull/38]
+* When transport is on UDP, transport isn't shutdown on network change. (#38)[https://github.com/VoIPGRID/VialerSIPLib/pull/38]
+* Proximity sensor is activated when the call is active (#41)[https://github.com/VoIPGRID/VialerSIPLib/pull/41]
+
+
+## [2.1.0](https://github.com/VoIPGRID/VialerSIPLib/tree/2.1.0) (01/09/2017)
+Released on Monday, January 9, 2017.
+
+#### Added
+* Bluetooth routes can be changed (#26)[https://github.com/VoIPGRID/VialerSIPLib/pull/26]
+
+
+#### Fixed
+* Updated documentation (#24)[https://github.com/VoIPGRID/VialerSIPLib/pull/24]
+* Start network monitoring only when there is a call active (#34)[https://github.com/VoIPGRID/VialerSIPLib/pull/34]
+* Network changes are delayed a little to prevent multiple register attempts (8c4c96a)[https://github.com/VoIPGRID/VialerSIPLib/commit/8c4c96ac76c4535abf47629dfe23b2c74a6498c7]
+
 
 ## [2.0.1](https://github.com/VoIPGRID/VialerSIPLib/tree/2.0.1) (12/09/2016)
 Released on Friday, December 9, 2016.